# openapi/openapi.lite.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Hatch CRM API (Lite)",
    "description": "Accounts, Opportunities, Cases, Files, Reporting",
    "version": "0.1.0"
  },
...
  "paths": {
...
    "/cases": {
      "get": {
        "tags": ["Cases"],
        "operationId": "listCases",
        "summary": "List cases",
        "parameters": [
          { "name": "q", "in": "query", "required": false, "schema": { "type": "string" } },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["New", "Working", "Escalated", "Resolved", "Closed"]
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": ["Low", "Medium", "High", "Urgent"]
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": { "type": "integer", "minimum": 1, "maximum": 200 }
          },
          { "name": "cursor", "in": "query", "required": false, "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "Case list",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseListResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "post": {
        "tags": ["Cases"],
        "operationId": "createCase",
        "summary": "Create case",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/CreateCase" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created case",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
    "/cases/{id}": {
      "parameters": [
        { "name": "id", "in": "path", "required": true, "schema": { "type": "string" } }
      ],
      "get": {
        "tags": ["Cases"],
        "operationId": "getCase",
        "summary": "Get case",
        "responses": {
          "200": {
            "description": "Case",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseResponse" }
              }
            }
          },
          "404": { "description": "Not found" }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "patch": {
        "tags": ["Cases"],
        "operationId": "updateCase",
        "summary": "Update case",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": { "$ref": "#/components/schemas/UpdateCase" }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated case",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/CaseResponse" }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      },
      "delete": {
        "tags": ["Cases"],
        "operationId": "deleteCase",
        "summary": "Delete case",
        "responses": {
          "200": {
            "description": "Deleted identifier",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "id": { "type": "string", "description": "Deleted case identifier" }
                  }
                }
              }
            }
          }
        },
        "security": [{ "bearerAuth": [] }]
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
...
      "CreateCase": {
        "type": "object",
        "properties": {
          "subject": { "type": "string" },
          "status": { "type": "string", "enum": ["New", "Working", "Escalated", "Resolved", "Closed"] },
          "priority": { "type": "string", "enum": ["Low", "Medium", "High", "Urgent"] },
          "origin": { "type": "string", "enum": ["Email", "Phone", "Web", "Other"] },
          "description": { "type": "string" },
          "accountId": { "type": "string" },
          "contactId": { "type": "string" }
        },
        "required": ["subject"],
        "additionalProperties": false
      },
      "UpdateCase": {
        "type": "object",
        "properties": {
          "subject": { "type": "string" },
          "status": { "type": "string", "enum": ["New", "Working", "Escalated", "Resolved", "Closed"] },
          "priority": { "type": "string", "enum": ["Low", "Medium", "High", "Urgent"] },
          "origin": { "type": "string", "enum": ["Email", "Phone", "Web", "Other"] },
          "description": { "type": "string" },
          "accountId": { "type": "string" },
          "contactId": { "type": "string" }
        },
        "additionalProperties": false
      },
      "CaseAccountSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": ["string", "null"] }
        },
        "required": ["id"],
        "additionalProperties": false
      },
      "CaseContactSummary": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": ["string", "null"] },
          "email": { "type": ["string", "null"], "format": "email" }
        },
        "required": ["id"],
        "additionalProperties": false
      },
      "CaseResponse": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "subject": { "type": ["string", "null"] },
          "status": { "type": ["string", "null"], "enum": ["New", "Working", "Escalated", "Resolved", "Closed"] },
          "priority": { "type": ["string", "null"], "enum": ["Low", "Medium", "High", "Urgent"] },
          "origin": { "type": ["string", "null"], "enum": ["Email", "Phone", "Web", "Other"] },
          "description": { "type": ["string", "null"] },
          "accountId": { "type": ["string", "null"] },
          "contactId": { "type": ["string", "null"] },
          "ownerId": { "type": ["string", "null"] },
          "createdAt": { "type": ["string", "null"], "format": "date-time" },
          "updatedAt": { "type": ["string", "null"], "format": "date-time" },
          "account": { "$ref": "#/components/schemas/CaseAccountSummary" },
          "contact": { "$ref": "#/components/schemas/CaseContactSummary" }
        },
        "required": ["id"],
        "additionalProperties": false
      },
      "CaseListResponse": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/CaseResponse" }
          },
          "nextCursor": { "type": ["string", "null"] }
        },
        "required": ["items"],
        "additionalProperties": false
      },
...
    }
  }
}